+++ cmd/list.go
@@ -1,3 +1,0 @@
-package cmd
-
-// Your code here