+++ templates/nvim/init.lua
@@ -1,2 +1,0 @@
-require("workspace.core")
-require("workspace.lazy")