+++ templates/nvim/lua/workspace/core/init.lua
@@ -1,3 +1,0 @@
-require("workspace.core.options")
-require("workspace.core.keymaps")
-require("workspace.core.autocmds")